+++ tests/bluetooth/bsim_bt/bsim_test_eatt/CMakeLists.txt
@@ -14,8 +14,9 @@ project(bsim_test_host)
 
 
 target_sources(app PRIVATE
-  src/main.c
+  src/common.c
   src/main_collision.c
+  src/main.c
 )
